Fix broken config options introduced in commit 1cc0032e4f
Commit 1cc0032e4f introduced two new config options: bind_ip and
bind_port. Unfortunately, it makes no sense to bind to a specific port
as only one user will be able to connect at a time (when a second user
tries to use iris, the (host, port) tuple will already be in use).
Additionally, the port was never converted from a str to an int, which
caused an exception if bind_ip was defined. Remove the bind_port option
entirely.

Since the bind_ip option has nothing to do with the actual execution of
the iris daemon and instead deals with connecting to the IRC server,
move it from the execution section to the irc section.