Support IPv6 in DNS lookup code
While Iris does not listen for IPv6 clients at the moment, it can be sat
behind a webserver that does, and if that server passes an appropriate
Forwarded-For header, the DNS lookup code in Iris won't work properly,
because it assumes IPv4 addresses & an in-addr.arpa. suffix for PTR
queries. This rectifies that.

Reported-by: Aaron Jones <aaronmdjones@gmail.com>
Authored-by: miniCruzer <samuelhoffman2@gmail.com>
Authored-by: Aaron Jones <aaronmdjones@gmail.com>